backend-api/requirements.txt
absl-py==2.1.0
accelerate==0.22.0
aiohttp==3.9.1
aiosignal==1.3.1
alabaster==0.7.16
altair==5.2.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autobahn==19.11.2
Automat==22.10.0
awadb==0.3.10
awsebcli==3.20.10
Babel==2.15.0
beautifulsoup4==4.12.3
black==24.4.2
blessed==1.20.0
blinker==1.7.0
botocore==1.31.85
CacheControl==0.14.0
cachetools==5.3.2
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.3
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
dataclasses-json==0.5.14
decorator==5.1.1
dnspython==2.6.1
docopt==0.6.2
docutils==0.20.1
einops==0.6.1
executing==2.0.1
faiss-cpu==1.7.4
fastembed==0.2.2
filelock==3.12.3
fire==0.6.0
firebase-admin==6.5.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flatbuffers==23.5.26
fonttools==4.51.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.130.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
gpt4all==2.1.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.5.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.23.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.34.1
imagesize==1.4.1
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==7.0.1
incremental==22.10.0
indic-nlp-library-IT2 @ git+https://github.com/VarunGumma/indic_nlp_library@601521e05ed0ed8f2165ac317a47d186e25b6f0d
-e git+https://github.com/VarunGumma/IndicTransTokenizer@0e68fb5872f4d821578a5252f90ad43c9649370f#egg=IndicTransTokenizer
iopath==0.1.10
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
Keras==2.3.0
Keras-Applications==1.0.8
keras-layer-normalization==0.16.0
Keras-Preprocessing==1.1.2
kiteconnect==5.0.1
kiwisolver==1.4.5
labelme2coco==0.2.6
langchain==0.0.354
langchain-community==0.0.13
langchain-core==0.1.11
langchain-google-genai==0.0.6
langchainplus-sdk==0.0.20
langsmith==0.0.81
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
Morfessor==2.0.6
mosestokenizer==1.2.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.1
nltk==3.8.1
numexpr==2.8.8
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.17.1
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openfile==0.0.7
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.4
parso==0.8.4
pathspec==0.10.1
peewee==3.17.1
pexpect==4.9.0
Pillow==10.0.0
pixellib==0.7.1
platformdirs==4.2.1
portalocker==2.8.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.3
sahi==0.11.15
scikit-image==0.23.2
scipy==1.13.0
semantic-version==2.8.5
sentencepiece==0.2.0
service-identity==24.1.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.25
st-theme==1.2.2
stack-data==0.6.3
streamlit==1.30.0
streamlit-navigation-bar==3.1.2
sympy==1.12
TA-Lib==0.4.28
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.15.0
termcolor==1.1.0
terminaltables==3.1.10
tifffile==2024.4.24
tokenizers==0.19.1
toml==0.10.2
toolwrapper==2.1.0
toolz==0.12.0
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.41.1
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.4
tzlocal==5.2
uctools==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
validators==0.22.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xmljson==0.2.1
yacs==0.1.8
yarl==1.9.4
yfinance==0.2.37
zipp==3.17.0
zope.interface==6.2
